Move City.java to dto package
